% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{active_run}
\alias{active_run}
\alias{mlflow_active_run}
\alias{mlflow_set_active_run}
\title{Active Run}
\usage{
mlflow_active_run()

mlflow_set_active_run(run)
}
\arguments{
\item{run}{The run object to make active.}
}
\description{
Retrieves or sets the active run.
}
