% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-utils.R
\name{build_context_tags_from_databricks_job_info}
\alias{build_context_tags_from_databricks_job_info}
\title{Get information from a Databricks job execution context}
\usage{
build_context_tags_from_databricks_job_info(job_info)
}
\arguments{
\item{job_info}{The job-related metadata from a running Databricks job}
}
\value{
A list of tags to be set by the run context when creating MLflow runs in the
current Databricks Job environment
}
\description{
Parses the data from a job execution context when running on Databricks in a non-interactive
mode. This function extracts relevant data that MLflow needs in order to properly utilize the
MLflow APIs from this context.
}
