% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-utils.R
\name{build_context_tags_from_databricks_notebook_info}
\alias{build_context_tags_from_databricks_notebook_info}
\title{Get information from Databricks Notebook environment}
\usage{
build_context_tags_from_databricks_notebook_info(notebook_info)
}
\arguments{
\item{notebook_info}{The configuration data from the Databricks Notebook environment}
}
\value{
A list of tags to be set by the run context when creating MLflow runs in the
current Databricks Notebook environment
}
\description{
Retrieves the notebook id, path, url, name, version, and type from the Databricks Notebook
execution environment and sets them to a list to be used for setting the configured environment
for executing an MLflow run in R from Databricks.
}
