% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-utils.R
\name{cast_choice}
\alias{cast_choice}
\title{Cast to choice (validate against allowed values)}
\usage{
cast_choice(x, choices, allow_null = FALSE)
}
\arguments{
\item{x}{Value to cast}

\item{choices}{Valid choices}

\item{allow_null}{Whether to allow NULL values}
}
\description{
Cast to choice (validate against allowed values)
}
