% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-utils.R
\name{cast_nullable_scalar_double}
\alias{cast_nullable_scalar_double}
\title{Cast to nullable scalar double}
\usage{
cast_nullable_scalar_double(x)
}
\arguments{
\item{x}{Value to cast}
}
\description{
Cast to nullable scalar double
}
