% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-utils.R
\name{cast_nullable_scalar_integer}
\alias{cast_nullable_scalar_integer}
\title{Cast to nullable scalar integer}
\usage{
cast_nullable_scalar_integer(x)
}
\arguments{
\item{x}{Value to cast}
}
\description{
Cast to nullable scalar integer
}
