% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-utils.R
\name{cast_scalar_double}
\alias{cast_scalar_double}
\title{Cast to scalar double}
\usage{
cast_scalar_double(x, allow_na = FALSE)
}
\arguments{
\item{x}{Value to cast}

\item{allow_na}{Whether to allow NA values}
}
\description{
Cast to scalar double
}
