% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-utils.R
\name{cast_string}
\alias{cast_string}
\title{Cast to string}
\usage{
cast_string(x, allow_na = FALSE)
}
\arguments{
\item{x}{Value to cast}

\item{allow_na}{Whether to allow NA values}
}
\description{
Cast to string
}
