% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-utils.R
\name{copy_project}
\alias{copy_project}
\title{Copy MLflow project directory for Docker deployment}
\usage{
copy_project(src, dst)
}
\arguments{
\item{src}{source directory}

\item{dst}{destination directory}
}
\description{
Utility function for copying MLflow project directory for Docker deployment,
which will copy everything from the source directory except for paths
matching patterns defined in the .dockerignore file.
It assumes MLflow is accessible as a local directory.
}
\keyword{internal}
