% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-utils.R
\name{.build_docker_image}
\alias{.build_docker_image}
\title{Build a MLflow Docker image}
\usage{
.build_docker_image(
  image_name,
  entry_point,
  mlflow_home = NULL,
  mlflow_version = utils::packageVersion("mlflow"),
  custom_setup_steps_hook = NULL
)
}
\arguments{
\item{image_name}{Name of the image}

\item{entry_point}{String containing ENTRYPOINT directive for the docker
image}

\item{mlflow_home}{(Optional) Path to local copy of MLflow. If unspecified,
then the image will be built with the version of MLflow specified by
`mlflow_version`.}

\item{mlflow_version}{(Optional) Ignored if `mlflow_home` is specified,
otherwise build the image with the  version of MLflow specified by
`mlflow_version` (default: `packageVersion("mlflow")`)}

\item{custom_setup_steps_hook}{(Optional) single-argument function accepting
the dockerfile context directory as input and returning additional
Dockerfile commands to run during the image build step as output.}
}
\description{
Utility function for building a Docker image for serving a MLflow model
}
\keyword{internal}
