% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tidymodels.R
\name{get_parsnip_dependencies}
\alias{get_parsnip_dependencies}
\title{Determine the package dependencies for a parsnip model}
\usage{
get_parsnip_dependencies(model, engine)
}
\arguments{
\item{model}{The type of model, eg. "linear_reg". This is stored as a class
of the model object, and is also recorded in the saved mlflow artefact.}

\item{engine}{The engine for the model, as understood by parsnip, eg. "lm".}
}
\value{
A character vector of required packages
}
\description{
Determine the package dependencies for a parsnip model
}
\keyword{internal}
