% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_mlflow}
\alias{install_mlflow}
\title{Install MLflow}
\usage{
install_mlflow()
}
\description{
Installs auxiliary dependencies of MLflow (e.g. the MLflow CLI). As a one-time setup step, you
must run install_mlflow() to install these dependencies before calling other MLflow APIs.
}
\details{
install_mlflow() requires Python and Conda to be installed.
See \url{https://www.python.org/getit/} and \url{https://docs.conda.io/projects/conda/en/latest/user-guide/install/}.
}
\examples{
\dontrun{
library(mlflow)
install_mlflow()
}

}
