% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-artifacts.R
\name{log_artifact_template}
\alias{log_artifact_template}
\title{Log an artifact with a transformer}
\usage{
log_artifact_template(transformer, artifact_file, ...)
}
\arguments{
\item{transformer}{A function that writes an object to disk.}

\item{artifact_file}{The run-relative artifact file path in
posixpath format to which the text is saved (e.g. "dir/file.txt").}

\item{...}{Passed to \code{transformer}. Must include a named argument
that specifies the R object to write. Because the name is not the same
for every transformer, this is passed through \code{...} and is not
a hardcoded argument in this function.}
}
\description{
These helper functions provide a mechanism to log artifacts in the run by fist
copying artifacts into a temp dir and then logging them to mlflow with
\code{\link[=mlflow_log_artifact]{mlflow_log_artifact()}}.
}
