% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcflow-package.R
\docType{package}
\name{qcflow-package}
\alias{qcflow}
\alias{qcflow-package}
\title{qcflow: Interface to 'QCFlow'}
\description{
R interface to 'QCFlow', open source platform for the complete machine learning life cycle, see <https://qcflow.org/>. This package supports installing 'QCFlow', tracking experiments, creating and running projects, and saving and serving models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qcflow/qcflow}
  \item Report bugs at \url{https://github.com/qcflow/qcflow/issues}
}

}
\author{
\strong{Maintainer}: Matei Zaharia \email{matei@databricks.com}

Authors:
\itemize{
  \item Javier Luraschi \email{jluraschi@gmail.com}
  \item Kevin Kuo \email{kevin.kuo@rstudio.com} (\href{https://orcid.org/0000-0001-7803-7901}{ORCID})
}

Other contributors:
\itemize{
  \item  RStudio [copyright holder]
}

}
\keyword{internal}
