% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_active_experiment}
\alias{mlflow_active_experiment}
\title{Active Experiment}
\usage{
mlflow_active_experiment()
}
\description{
Retrieves the active experiment which got activated using \code{mlflow_experiment()} or
the \code{MLFLOW_EXPERIMENT_ID} environemnt variable.
}
