% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tidymodels.R
\name{mlflow_autolog_params}
\alias{mlflow_autolog_params}
\title{Automatically log the parameters of a model}
\usage{
mlflow_autolog_params(x, ...)
}
\arguments{
\item{x}{A model of supported type.}

\item{...}{Further arguments passed to methods.}
}
\value{
The input model, x.
}
\description{
Currently only implemented for \verb{parsnip} models and \verb{workflow}s.
}
