% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_client_create_experiment}
\alias{mlflow_client_create_experiment}
\title{Create Experiment - Tracking Client}
\usage{
mlflow_client_create_experiment(client, name, artifact_location = NULL)
}
\arguments{
\item{client}{}

\item{name}{The name of the experiment to create.}

\item{artifact_location}{Location where all artifacts for this experiment are stored. If
not provided, the remote server will select an appropriate default.}
}
\description{
Creates an MLflow experiment.
}
