% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_client_delete_experiment}
\alias{mlflow_client_delete_experiment}
\title{Delete Experiment}
\usage{
mlflow_client_delete_experiment(client, experiment_id)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{experiment_id}{ID of the associated experiment This field is required.}
}
\description{
Mark an experiment and associated runs, params, metrics, … etc for deletion. If the
  experiment uses FileStore, artifacts associated with experiment are also deleted.
}
