% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_client_get_experiment}
\alias{mlflow_client_get_experiment}
\title{Get Experiment}
\usage{
mlflow_client_get_experiment(client, experiment_id)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{experiment_id}{Identifer to get an experiment.}
}
\description{
Get meta data for experiment and a list of runs for this experiment.
}
