% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_client_get_experiment_by_name}
\alias{mlflow_client_get_experiment_by_name}
\title{Get Experiment by Name}
\usage{
mlflow_client_get_experiment_by_name(client, name)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{name}{The experiment name.}
}
\description{
Get meta data for experiment by name.
}
