% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_client_get_run}
\alias{mlflow_client_get_run}
\title{Get Run}
\usage{
mlflow_client_get_run(client, run_uuid)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{run_uuid}{Unique ID for the run.}
}
\description{
Get meta data, params, tags, and metrics for run. Only last logged value for each metric is returned.
}
