% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_client_log_param}
\alias{mlflow_client_log_param}
\title{Log Parameter}
\usage{
mlflow_client_log_param(client, run_uuid, key, value)
}
\arguments{
\item{key}{Name of the parameter.}

\item{value}{String value of the parameter.}
}
\description{
API to log a parameter used for this run. Examples are params and hyperparams
  used for ML training, or constant dates and values used in an ETL pipeline.
  A params is a STRING key-value pair. For a run, a single parameter is allowed
  to be logged only once.
}
