% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_client_set_terminated}
\alias{mlflow_client_set_terminated}
\title{Terminate a Run}
\usage{
mlflow_client_set_terminated(client, run_id, status = c("FINISHED",
  "SCHEDULED", "FAILED", "KILLED"), end_time = NULL)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{run_id}{Unique identifier for the run.}

\item{status}{Updated status of the run. Defaults to `FINISHED`.}

\item{end_time}{Unix timestamp of when the run ended in milliseconds.}
}
\description{
Terminate a Run
}
