% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{mlflow_connect}
\alias{mlflow_connect}
\title{Connect to MLflow}
\usage{
mlflow_connect(url = NULL, store = NULL, artifacts = NULL)
}
\arguments{
\item{url}{Optional URL to the remote MLflow server; otherwise,
will launch and connect local instance.}

\item{store}{The root of the backing file store for
experiment and run data. Defaults to \code{./mlruns}.}

\item{artifacts}{Local or S3 URI to store artifacts in, for
newly created experiments. Note that this flag
does not impact already-created experiments.
Defaults to a location inside \code{store}.}
}
\description{
Connect to local or remote MLflow instance.
}
