% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{mlflow_create_model_version}
\alias{mlflow_create_model_version}
\title{Create a model version}
\usage{
mlflow_create_model_version(
  name,
  source,
  run_id = NULL,
  tags = NULL,
  run_link = NULL,
  description = NULL,
  client = NULL
)
}
\arguments{
\item{name}{Register model under this name.}

\item{source}{URI indicating the location of the model artifacts.}

\item{run_id}{MLflow run ID for correlation, if `source` was generated
by an experiment run in MLflow Tracking.}

\item{tags}{Additional metadata.}

\item{run_link}{MLflow run link - This is the exact link of the run that
generated this model version.}

\item{description}{Description for model version.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Create a model version
}
