% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{mlflow_create_registered_model}
\alias{mlflow_create_registered_model}
\title{Create registered model}
\usage{
mlflow_create_registered_model(
  name,
  tags = NULL,
  description = NULL,
  client = NULL
)
}
\arguments{
\item{name}{The name of the model to create.}

\item{tags}{Additional metadata for the registered model (Optional).}

\item{description}{Description for the registered model (Optional).}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Creates a new registered model in the model registry
}
