% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_delete_experiment}
\alias{mlflow_delete_experiment}
\title{Delete Experiment}
\usage{
mlflow_delete_experiment(experiment_id, client)
}
\arguments{
\item{experiment_id}{ID of the associated experiment This field is required.}

\item{client}{An `mlflow_client` object.}
}
\description{
Mark an experiment and associated runs, params, metrics, … etc for deletion. If the
  experiment uses FileStore, artifacts associated with experiment are also deleted.
}
