% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{mlflow_delete_registered_model}
\alias{mlflow_delete_registered_model}
\title{Delete registered model}
\usage{
mlflow_delete_registered_model(name, client = NULL)
}
\arguments{
\item{name}{The name of the model to delete}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Deletes an existing registered model by name
}
