% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{mlflow_disconnect}
\alias{mlflow_disconnect}
\title{Disconnect from MLflow}
\usage{
mlflow_disconnect(mc)
}
\arguments{
\item{mc}{The MLflow connection created using \code{mlflow_connect()}.}
}
\description{
Disconnects from a local or remote MLflow instance.
}
