% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-fluent.R
\name{mlflow_end_run}
\alias{mlflow_end_run}
\title{End a Run}
\usage{
mlflow_end_run(status = c("FINISHED", "SCHEDULED", "FAILED", "KILLED"))
}
\arguments{
\item{status}{Updated status of the run. Defaults to `FINISHED`.}
}
\description{
End an active MLflow run (if there is one).
}
