% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/experiment.R
\name{mlflow_experiment}
\alias{mlflow_experiment}
\title{Active Experiment}
\usage{
mlflow_experiment(name)
}
\arguments{
\item{name}{The name of the experiment to create.}
}
\description{
Creates an MLflow experiment and makes it active.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()

# activates experiment
mlflow_experiment("My Experiment")

# activates experiment in remote MLflow server
mlflow_tracking_url("http://tracking-server:5000")
mlflow_experiment("My Experiment")
}

}
