% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_experiment_create}
\alias{mlflow_experiment_create}
\title{Create Experiment}
\usage{
mlflow_experiment_create(name)
}
\arguments{
\item{name}{The name of the experiment to create.}
}
\description{
Creates an MLflow experiment.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()

# create local experiment
mlflow_experiment_create("My Experiment")

# create experiment in remote MLflow server
mlflow_tracking_url("http://tracking-server:5000")
mlflow_experiments_create("My Experiment")
}

}
