% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_experiments}
\alias{mlflow_experiments}
\title{List Experiments}
\usage{
mlflow_experiments(mc)
}
\arguments{
\item{mc}{The MLflow connection created using \code{mlflow_connect()}.}
}
\description{
Retrieves MLflow experiments as a data frame.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()

mc <- mlflow_connect()
mlflow_experiments(mc)
}

}
