% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mlflow_flavor}
\alias{mlflow_flavor}
\title{Create an MLflow Flavor Object}
\usage{
mlflow_flavor(flavor)
}
\arguments{
\item{flavor}{The name of the flavor.}
}
\description{
This function creates an `mlflow_flavor` object that can be used to dispatch
  the `mlflow_load_flavor()` method.
}
\keyword{internal}
