% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_get_experiment_by_name}
\alias{mlflow_get_experiment_by_name}
\title{Get Experiment by Name}
\usage{
mlflow_get_experiment_by_name(name, client)
}
\arguments{
\item{name}{The experiment name.}

\item{client}{An `mlflow_client` object.}
}
\description{
Get meta data for experiment by name.
}
