% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{mlflow_get_latest_versions}
\alias{mlflow_get_latest_versions}
\title{Get latest model versions}
\usage{
mlflow_get_latest_versions(name, stages = list(), client = NULL)
}
\arguments{
\item{name}{Name of the model.}

\item{stages}{A list of desired stages. If the input list is NULL, return
latest versions for ALL_STAGES.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Retrieves a list of the latest model versions for a given model.
}
