% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/experiment.R
\name{mlflow_get_metric}
\alias{mlflow_get_metric}
\title{Get Metric}
\usage{
mlflow_get_metric(run_uuid, metric_key)
}
\arguments{
\item{run_uuid}{Unique ID for the run for which metric is recorded.}

\item{metric_key}{Name of the metric.}
}
\description{
API to retrieve the logged value for a metric during a run. For a run, if this
  metric is logged more than once, this API will retrieve only the latest value logged.
}
