% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_get_param}
\alias{mlflow_get_param}
\title{Get Param}
\usage{
mlflow_get_param(param_name, client, run_id)
}
\arguments{
\item{param_name}{Name of the param. This field is required.}

\item{client}{An `mlflow_client` object.}

\item{run_id}{Run ID.}
}
\value{
The param name and value as a data frame.
}
\description{
Get a param value.
}
