% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{mlflow_list_registered_models}
\alias{mlflow_list_registered_models}
\title{List registered models}
\usage{
mlflow_list_registered_models(
  max_results = 100,
  page_token = NULL,
  client = NULL
)
}
\arguments{
\item{max_results}{Maximum number of registered models to retrieve.}

\item{page_token}{Pagination token to go to the next page based on a
previous query.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Retrieves a list of registered models.
}
