% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tidymodels.R, R/model.R
\name{mlflow_load_flavor.mlflow_flavor_workflow}
\alias{mlflow_load_flavor.mlflow_flavor_workflow}
\alias{mlflow_load_flavor.mlflow_flavor_parsnip}
\alias{mlflow_load_flavor}
\title{Load MLflow Model Flavor}
\usage{
\method{mlflow_load_flavor}{mlflow_flavor_workflow}(flavor, model_path)

\method{mlflow_load_flavor}{mlflow_flavor_parsnip}(flavor, model_path)

mlflow_load_flavor(flavor, model_path)
}
\arguments{
\item{flavor}{An MLflow flavor object loaded by \link[mlflow]{mlflow_load_model}, with class
loaded from the flavor field in an MLmodel file.}

\item{model_path}{The path to the MLflow model wrapped in the correct
class.}
}
\description{
Loads an MLflow model using a specific flavor. This method is called internally by
\link[mlflow]{mlflow_load_model}, but is exposed for package authors to extend the supported
MLflow models. See https://mlflow.org/docs/latest/models.html#storage-format for more
info on MLflow model flavors.
}
