% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_log}
\alias{mlflow_log}
\title{Log to MLflow}
\usage{
mlflow_log(name, value)
}
\arguments{
\item{name}{The name to identify this log entry.}

\item{value}{The value to log into this entry.}
}
\description{
Logs a value to MLflow for the active run.
}
