% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_log_artifact.mlflow_client}
\alias{mlflow_log_artifact.mlflow_client}
\title{Log Artifact}
\usage{
mlflow_log_artifact.mlflow_client(client, run_id, path,
  artifact_path = NULL)
}
\arguments{
\item{client}{An `mlflow_client` object. See details.}

\item{run_id}{The run associated with this artifact.}

\item{path}{The file or directory to log as an artifact.}

\item{artifact_path}{Destination path within the run’s artifact URI.}

\item{...}{Additional arguments; currently not used.}
}
\description{
Logs an specific file or directory as an artifact.
}
\details{
For functions where `client` is optional, providing it will invoke the
  Tracking Service API, allowing the user to specify the service where
  the operation is executed. When `client` isn't specified, it defaults to
  the service set by `mlflow_set_tracking_uri()`.

When logging to Amazon S3, ensure that the user has a proper policy
attach to it, for instance:

\code{
{
"Version": "2012-10-17",
"Statement": [
  {
    "Sid": "VisualEditor0",
    "Effect": "Allow",
    "Action": [
      "s3:PutObject",
      "s3:GetObject",
      "s3:ListBucket",
      "s3:GetBucketLocation"
      ],
    "Resource": [
      "arn:aws:s3:::mlflow-test/*",
      "arn:aws:s3:::mlflow-test"
      ]
  }
  ]
}
}

Additionally, at least the \code{AWS_ACCESS_KEY_ID} and \code{AWS_SECRET_ACCESS_KEY}
environment variables must be set to the corresponding key and secrets provided
by Amazon IAM.
}
