% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-artifacts.R
\name{mlflow_log_text}
\alias{mlflow_log_text}
\title{Log text as an artifact.}
\usage{
mlflow_log_text(text, artifact_file)
}
\arguments{
\item{text}{String containing text to log.}

\item{artifact_file}{The run-relative artifact file path in
posixpath format to which the text is saved (e.g. "dir/file.txt").}
}
\description{
Log text as an artifact.
}
\examples{
\dontrun{
library(mlflow)
with(mlflow_start_run(), {
  # Log text to a file under the run's root artifact directory
  mlflow_log_text("text1", "file1.txt")
  # Log text in a subdirectory of the run's root artifact directory
  mlflow_log_text("text2", "dir/file2.txt")
  # Log HTML text
  mlflow_log_text("<h1>header</h1>", "index.html")
})
}
}
