% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-flavor.R
\name{mlflow_predict_flavor}
\alias{mlflow_predict_flavor}
\title{Predict over MLflow Model Flavor}
\usage{
mlflow_predict_flavor(model, data)
}
\arguments{
\item{model}{The loaded MLflow model flavor.}

\item{data}{A data frame to perform scoring.}
}
\description{
Performs prediction over a model loaded using
\code{mlflow_load_model()}, to be used by package authors
to extend the supported MLflow models.
}
