% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-generics.R
\name{mlflow_rename_experiment}
\alias{mlflow_rename_experiment}
\title{Rename Experiment}
\usage{
mlflow_rename_experiment(experiment_id, new_name, client = NULL)
}
\arguments{
\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{new_name}{The experiment’s name will be changed to this. The new name must be unique.}
}
\description{
Renames an experiment.
}
