% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{mlflow_restore_run}
\alias{mlflow_restore_run}
\title{Restore a Run}
\usage{
mlflow_restore_run(run_id, client = NULL)
}
\arguments{
\item{run_id}{Run ID.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Restore a Run
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
