% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-snapshot.R
\name{mlflow_restore_snapshot}
\alias{mlflow_restore_snapshot}
\title{Restore Snapshot}
\usage{
mlflow_restore_snapshot()
}
\description{
Restores a snapshot of all dependencies required to run the files in the
current directory.
}
