% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{mlflow_run}
\alias{mlflow_run}
\title{Run in MLflow}
\usage{
mlflow_run(x, params = NULL)
}
\arguments{
\item{x}{A directory or an R script.}

\item{params}{A list of parameters.}
}
\description{
Runs the given file, expression or function within
the context of an MLflow run.
}
