% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-keras.R
\name{mlflow_save_flavor.keras.engine.training.Model}
\alias{mlflow_save_flavor.keras.engine.training.Model}
\title{importFrom mlflow create_conda_env, create_pyfunc_conf}
\usage{
\method{mlflow_save_flavor}{keras.engine.training.Model}(x,
  path = "model", r_dependencies = NULL, conda_env = NULL)
}
\description{
importFrom mlflow create_conda_env, create_pyfunc_conf
}
