% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tidymodels.R
\name{mlflow_save_model.model_spec}
\alias{mlflow_save_model.model_spec}
\title{Attempt to save a parsnip model that is yet to be fitted}
\usage{
\method{mlflow_save_model}{model_spec}(model, path, model_spec = list(), ...)
}
\arguments{
\item{model}{The model that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}

\item{model_spec}{MLflow model config this model flavor is being added to.}

\item{...}{Optional additional arguments.}
}
\description{
This method exists only to return an error message if the user attempts to
save a parsnip model which is yet to be fit. These have the "model_spec"
class, whereas a fitted parsnip models have the "model_fit" path.
}
\keyword{internal}
