% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/model.R
\name{mlflow_save_model}
\alias{mlflow_save_model}
\title{Save Model for MLflow}
\usage{
mlflow_save_model(f, path = "mlflow-model")
}
\arguments{
\item{f}{The serving function that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}
}
\description{
Saves model in MLflow's format that can later be used
for prediction and serving.
}
