% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{mlflow_script}
\alias{mlflow_script}
\title{Run Script}
\usage{
mlflow_script(script)
}
\arguments{
\item{script}{Path to the script to be run.}
}
\description{
Runs an R script in the context of 'MLflow', designed to be used by
tools like terminal or RStudio.
}
