% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_search_runs}
\alias{mlflow_search_runs}
\title{Search Runs}
\usage{
mlflow_search_runs(experiment_ids, anded_expressions)
}
\arguments{
\item{experiment_ids}{An array of integers.}

\item{anded_expressions}{An array of SearchExpression.}
}
\description{
Search for runs that satisfy expressions. Search expressions can use Metric and Param keys.
}
