% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-fluent.R
\name{mlflow_set_experiment}
\alias{mlflow_set_experiment}
\title{Set Experiment}
\usage{
mlflow_set_experiment(experiment_name)
}
\arguments{
\item{experiment_name}{Name of experiment to be activated.}
}
\description{
Set given experiment as active experiment. If experiment does not
 exist, create an experiment with provided name.
}
