% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{mlflow_set_registered_model_alias}
\alias{mlflow_set_registered_model_alias}
\title{Set a model alias}
\usage{
mlflow_set_registered_model_alias(name, alias, version, client = NULL, ...)
}
\arguments{
\item{name}{Name of the registered model.}

\item{alias}{Alias to set.}

\item{version}{Model version number.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}

\item{...}{Reserved for future options.}
}
\description{
Assigns an alias to a registered model version.
}
\details{
Load an aliased model with `mlflow_load_model("models:/<model_name>@<alias>")`.
}
\examples{
\dontrun{
mlflow_set_registered_model_alias("catalog.schema.model", "champion", 1)
mlflow_load_model("models:/catalog.schema.model@champion")
}
}
