% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-globals.R
\name{mlflow_set_registry_uri}
\alias{mlflow_set_registry_uri}
\title{Set Model Registry URI}
\usage{
mlflow_set_registry_uri(uri)
}
\arguments{
\item{uri}{The URI to the model registry service.}
}
\description{
Specifies the URI to the model registry service used for model registry operations.
}
\details{
For Databricks model registry, use `databricks-uc` or
  `databricks-uc://<profile>`. When the tracking URI is `databricks`,
  this is the default registry URI.
}
