% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow-fluent.R, R/tracking.R
\name{mlflow_set_tag.NULL}
\alias{mlflow_set_tag.NULL}
\alias{mlflow_set_tag}
\alias{mlflow_set_tag.mlflow_client}
\title{Set Tag}
\usage{
\method{mlflow_set_tag}{NULL}(key, value, client = NULL, ...)

mlflow_set_tag(key, value, client = NULL, ...)

\method{mlflow_set_tag}{mlflow_client}(key, value, client = NULL, run_id,
  ...)
}
\arguments{
\item{key}{Name of the tag. Maximum size is 255 bytes. This field is required.}

\item{value}{String value of the tag being logged. Maximum size is 500 bytes. This field is required.}

\item{run_id}{Run ID.}
}
\description{
Set a tag on a run. Tags are run metadata that can be updated during and
 after a run completes.
}
