% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-globals.R
\name{mlflow_set_tracking_uri}
\alias{mlflow_set_tracking_uri}
\title{Set Remote Tracking URI}
\usage{
mlflow_set_tracking_uri(uri)
}
\arguments{
\item{uri}{The URI to the remote MLflow server.}
}
\description{
Specifies the URI to the remote MLflow server that will be used
to track experiments.
}
