% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-snapshot.R
\name{mlflow_snapshot}
\alias{mlflow_snapshot}
\title{Create Dependency Snapshot}
\usage{
mlflow_snapshot()
}
\description{
Creates a snapshot of all dependencies required to run the files in the
current directory.
}
