% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_start_run}
\alias{mlflow_start_run}
\title{Start Run}
\usage{
mlflow_start_run(experiment_id = mlflow_active_experiment(), user_id = NULL,
  run_name = NULL, source_type = NULL, source_name = NULL,
  status = NULL, start_time = NULL, end_time = NULL,
  source_version = NULL, artifact_uri = NULL, entry_point_name = NULL,
  run_tags = NULL)
}
\arguments{
\item{experiment_id}{Unique identifier for the associated experiment.}

\item{user_id}{User ID or LDAP for the user executing the run.}

\item{run_name}{Human readable name for run.}

\item{source_type}{Originating source for this run. One of Notebook, Job, Project, Local or Unknown.}

\item{source_name}{String descriptor for source. For example, name or description of the notebook, or job name.}

\item{status}{Current status of the run. One of RUNNING, SCHEDULE, FINISHED, FAILED, KILLED.}

\item{start_time}{Unix timestamp of when the run started in milliseconds.}

\item{end_time}{Unix timestamp of when the run ended in milliseconds.}

\item{source_version}{Git version of the source code used to create run.}

\item{artifact_uri}{URI of the directory where artifacts should be uploaded This can be a local path (starting with “/”),
or a distributed file system (DFS) path, like s3://bucket/directory or dbfs:/my/directory. If not set, the local ./mlruns
directory will be chosen by default.}

\item{entry_point_name}{Name of the entry point for the run.}

\item{run_tags}{Additional metadata for run in key-value pairs.}
}
\description{
Starts a new run within an experiment, should be used within a \code{with} block.
}
\examples{
\dontrun{
with(mlflow_start_run(), {
  mlflow_log("test", 10)
})
}

}
