% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/remote.R
\name{mlflow_tracking_url}
\alias{mlflow_tracking_url}
\title{Remote Tracking URL}
\usage{
mlflow_tracking_url(url)
}
\arguments{
\item{url}{The URL to the remote MLflow server.}
}
\description{
Specifies the URL to the remote MLflow server that will be used
to track experiments.
}
