% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{mlflow_transition_model_version_stage}
\alias{mlflow_transition_model_version_stage}
\title{Transition ModelVersion Stage}
\usage{
mlflow_transition_model_version_stage(
  name,
  version,
  stage,
  archive_existing_versions = FALSE,
  client = NULL
)
}
\arguments{
\item{name}{Name of the registered model.}

\item{version}{Model version number.}

\item{stage}{Transition `model_version` to this stage.}

\item{archive_existing_versions}{(Optional)}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Transition a model version to a different stage.
}
