% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{mlflow_ui}
\alias{mlflow_ui}
\title{MLflow User Interface}
\usage{
mlflow_ui(mc)
}
\arguments{
\item{mc}{The MLflow connection created using \code{mlflow_connect()}.}
}
\description{
Launches MLflow user interface.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()

mc <- mlflow_connect()
mlflow_ui(mc)
}

}
