% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{mlflow_update_registered_model}
\alias{mlflow_update_registered_model}
\title{Update a registered model}
\usage{
mlflow_update_registered_model(name, description, client = NULL)
}
\arguments{
\item{name}{The name of the registered model.}

\item{description}{The updated description for this registered model.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Updates a model in the Model Registry.
}
