% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/experiment.R
\name{mlflow_update_run}
\alias{mlflow_update_run}
\title{Update Run}
\usage{
mlflow_update_run(run_uuid, status = c("FINISHED", "SCHEDULED", "FAILED",
  "KILLED"), end_time = NULL)
}
\arguments{
\item{run_uuid}{Unique identifier for the run.}

\item{status}{Updated status of the run. Defaults to `FINISHED`.}

\item{end_time}{Unix timestamp of when the run ended in milliseconds.}
}
\description{
Update Run
}
