% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mlflow_write_model_spec}
\alias{mlflow_write_model_spec}
\title{Write Model Specification}
\usage{
mlflow_write_model_spec(path, content)
}
\arguments{
\item{path}{Destination path where this MLflow compatible model
will be saved.}

\item{content}{The content to be saved to the MLmodel
specification.}
}
\description{
Provides support to extend new model flavors, by subclassing
\code{mlflow_save_model()} and performing a call to this
function to write the flavor specification.
}
