% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%||\%}
\alias{\%>\%}
\alias{crate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{carrier}{\code{\link[carrier]{crate}}}

  \item{purrr}{\code{\link[purrr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{\%||\%}}}
}}

