% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tidymodels.R
\name{require_parsnip_dependencies}
\alias{require_parsnip_dependencies}
\title{Check that the package dependencies for a parsnip model are satisfied}
\usage{
require_parsnip_dependencies(model, engine)
}
\arguments{
\item{model}{The type of model, eg. "linear_reg". This is stored as a class
of the model object, and is also recorded in the saved mlflow artefact.}

\item{engine}{The engine for the model, as understood by parsnip, eg. "lm".}
}
\value{
Invisibly returns a character vector of required packages
}
\description{
Check that the package dependencies for a parsnip model are satisfied
}
\keyword{internal}
