% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{uninstall_mlflow}
\alias{uninstall_mlflow}
\title{Uninstall MLflow}
\usage{
uninstall_mlflow()
}
\description{
Uninstalls MLflow by removing the Conda environment.
}
\examples{
\dontrun{
library(mlflow)
install_mlflow()
uninstall_mlflow()
}

}
