% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{active_experiment}
\alias{active_experiment}
\alias{mlflow_active_experiment}
\alias{mlflow_set_active_experiment}
\title{Active Experiment}
\usage{
mlflow_active_experiment()

mlflow_set_active_experiment(experiment_id)
}
\arguments{
\item{experiment_id}{Identifer to get an experiment.}
}
\description{
Retrieve or set the active experiment.
}
