% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow-package.R
\docType{package}
\name{mlflow-package}
\alias{mlflow}
\alias{mlflow-package}
\title{mlflow: Interface to 'MLflow'}
\description{
R interface to 'MLflow', open source platform for the complete machine learning life cycle, see <https://mlflow.org/>. This package supports installing 'MLflow', tracking experiments, creating and running projects, and saving and serving models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlflow/mlflow}
  \item Report bugs at \url{https://github.com/mlflow/mlflow/issues}
}

}
\author{
\strong{Maintainer}: Ben Wilson \email{benjamin.wilson@databricks.com}

Authors:
\itemize{
  \item Matei Zaharia \email{matei@databricks.com}
  \item Javier Luraschi \email{jluraschi@gmail.com}
  \item Kevin Kuo \email{kevin.kuo@rstudio.com} (\href{https://orcid.org/0000-0001-7803-7901}{ORCID})
}

Other contributors:
\itemize{
  \item  RStudio [copyright holder]
}

}
\keyword{internal}
