% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_active_experiment}
\alias{mlflow_active_experiment}
\title{Active Experiment}
\usage{
mlflow_active_experiment()
}
\description{
Retrieves the active experiment. An experiment is made active by calling
\code{mlflow_experiment()}.
}
