% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_client}
\alias{mlflow_client}
\title{Initialize an MLflow client}
\usage{
mlflow_client(tracking_uri = NULL)
}
\arguments{
\item{tracking_uri}{The tracking URI}
}
\description{
Initialize an MLflow client
}
\keyword{internal}
