% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_client}
\alias{mlflow_client}
\title{Initialize an MLflow Client}
\usage{
mlflow_client(tracking_uri = NULL)
}
\arguments{
\item{tracking_uri}{The tracking URI. If not provided, defaults to the service
set by `mlflow_set_tracking_uri()`.}
}
\description{
Initializes and returns an MLflow client that communicates with the tracking server or store
at the specified URI.
}
