% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client-methods.R
\name{mlflow_client_delete_experiment}
\alias{mlflow_client_delete_experiment}
\title{Delete Experiment}
\usage{
mlflow_client_delete_experiment(client, experiment_id)
}
\arguments{
\item{client}{(Optional) An `mlflow_client` object.}

\item{experiment_id}{ID of the associated experiment. This field is required.}
}
\description{
Marks an experiment and associated runs, params, metrics, etc. for deletion. If the
  experiment uses FileStore, artifacts associated with experiment are also deleted.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
\seealso{
Other tracking functions: \code{\link{mlflow_client_create_run}},
  \code{\link{mlflow_client_delete_run}},
  \code{\link{mlflow_client_download_artifacts}},
  \code{\link{mlflow_client_get_experiment_by_name}},
  \code{\link{mlflow_client_get_run}},
  \code{\link{mlflow_client_list_artifacts}},
  \code{\link{mlflow_client_log_artifact}},
  \code{\link{mlflow_client_log_param}},
  \code{\link{mlflow_client_restore_experiment}},
  \code{\link{mlflow_client_restore_run}},
  \code{\link{mlflow_client_set_tag}},
  \code{\link{mlflow_client_set_terminated}},
  \code{\link{mlflow_create_experiment}},
  \code{\link{mlflow_get_experiment}},
  \code{\link{mlflow_list_experiments}},
  \code{\link{mlflow_log_metric}}
}
\concept{tracking functions}
