% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_client_get_experiment}
\alias{mlflow_client_get_experiment}
\title{Get Experiment}
\usage{
mlflow_client_get_experiment(client, experiment_id)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{experiment_id}{Identifer to get an experiment.}
}
\description{
Get meta data for experiment and a list of runs for this experiment.
}
\details{
The Tracking Client family of functions require an MLflow client to be
  specified explicitly. These functions allow for greater control of where the
  operations take place in terms of services and runs, but are more verbose
  compared to the Fluent API.
}
\keyword{internal}
