% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_client_get_run}
\alias{mlflow_client_get_run}
\title{Get Run}
\usage{
mlflow_client_get_run(client, run_uuid)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{run_uuid}{Unique ID for the run.}
}
\description{
Get meta data, params, tags, and metrics for run. Only last logged value for each metric is returned.
}
\seealso{
Other Tracking client functions: \code{\link{mlflow_client_create_experiment}},
  \code{\link{mlflow_client_create_run}},
  \code{\link{mlflow_client_delete_experiment}},
  \code{\link{mlflow_client_delete_run}},
  \code{\link{mlflow_client_get_experiment_by_name}},
  \code{\link{mlflow_client_get_experiment}},
  \code{\link{mlflow_client_list_experiments}},
  \code{\link{mlflow_client_log_artifact}},
  \code{\link{mlflow_client_log_metric}},
  \code{\link{mlflow_client_log_param}},
  \code{\link{mlflow_client_restore_experiment}},
  \code{\link{mlflow_client_restore_run}},
  \code{\link{mlflow_client_set_tag}},
  \code{\link{mlflow_client_set_terminated}}
}
\concept{Tracking client functions}
\keyword{internal}
