% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_client_get_run}
\alias{mlflow_client_get_run}
\title{Get Run}
\usage{
mlflow_client_get_run(client, run_id)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{run_id}{Run ID.}
}
\description{
Gets metadata, params, tags, and metrics for a run. In the case where multiple metrics with the
same key are logged for the run, returns only the value with the latest timestamp. If there are
multiple values with the latest timestamp, returns the maximum of these values.
}
\details{
The Tracking Client family of functions require an MLflow client to be
  specified explicitly. These functions allow for greater control of where the
  operations take place in terms of services and runs, but are more verbose
  compared to the Fluent API.
}
\keyword{internal}
