% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_client_log_param}
\alias{mlflow_client_log_param}
\title{Log Parameter}
\usage{
mlflow_client_log_param(client, run_id, key, value)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{run_id}{Run ID.}

\item{key}{Name of the parameter.}

\item{value}{String value of the parameter.}
}
\description{
API to log a parameter used for this run. Examples are params and hyperparams
  used for ML training, or constant dates and values used in an ETL pipeline.
  A params is a STRING key-value pair. For a run, a single parameter is allowed
  to be logged only once.
}
\details{
The Tracking Client family of functions require an MLflow client to be
  specified explicitly. These functions allow for greater control of where the
  operations take place in terms of services and runs, but are more verbose
  compared to the Fluent API.
}
\keyword{internal}
