% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client-methods.R
\name{mlflow_client_set_terminated}
\alias{mlflow_client_set_terminated}
\title{Terminate a Run}
\usage{
mlflow_client_set_terminated(client, run_id, status = c("FINISHED",
  "SCHEDULED", "FAILED", "KILLED"), end_time = NULL)
}
\arguments{
\item{client}{(Optional) An `mlflow_client` object.}

\item{run_id}{Unique identifier for the run.}

\item{status}{Updated status of the run. Defaults to `FINISHED`.}

\item{end_time}{Unix timestamp of when the run ended in milliseconds.}

\item{run_id}{Run ID.}
}
\description{
Terminates a run.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
\seealso{
Other tracking functions: \code{\link{mlflow_client_create_run}},
  \code{\link{mlflow_client_delete_run}},
  \code{\link{mlflow_client_download_artifacts}},
  \code{\link{mlflow_client_get_experiment_by_name}},
  \code{\link{mlflow_client_get_run}},
  \code{\link{mlflow_client_list_artifacts}},
  \code{\link{mlflow_client_log_artifact}},
  \code{\link{mlflow_client_log_param}},
  \code{\link{mlflow_client_restore_experiment}},
  \code{\link{mlflow_client_restore_run}},
  \code{\link{mlflow_client_set_tag}},
  \code{\link{mlflow_create_experiment}},
  \code{\link{mlflow_delete_experiment}},
  \code{\link{mlflow_get_experiment}},
  \code{\link{mlflow_list_experiments}},
  \code{\link{mlflow_log_metric}}
}
\concept{tracking functions}
