% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_client_set_terminated}
\alias{mlflow_client_set_terminated}
\title{Terminate a Run}
\usage{
mlflow_client_set_terminated(client, run_id, status = c("FINISHED",
  "SCHEDULED", "FAILED", "KILLED"), end_time = NULL)
}
\arguments{
\item{client}{An `mlflow_client` object.}

\item{run_id}{Unique identifier for the run.}

\item{status}{Updated status of the run. Defaults to `FINISHED`.}

\item{end_time}{Unix timestamp of when the run ended in milliseconds.}

\item{run_id}{Run ID.}
}
\description{
Terminate a Run
}
\details{
The Tracking Client family of functions require an MLflow client to be
  specified explicitly. These functions allow for greater control of where the
  operations take place in terms of services and runs, but are more verbose
  compared to the Fluent API.
}
\keyword{internal}
