% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{mlflow_connect}
\alias{mlflow_connect}
\title{Connect to MLflow}
\usage{
mlflow_connect(url = NULL, ...)
}
\arguments{
\item{url}{Optional URL to the remote MLflow server. If not specified,
will launch and connect to a local instance listening on a random port.}

\item{...}{Options arguments passed to `mlflow_server()`.}
}
\description{
Connect to local or remote MLflow instance.
}
