% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{mlflow_connect}
\alias{mlflow_connect}
\title{Connect to MLflow}
\usage{
mlflow_connect(x = NULL, activate = TRUE, ...)
}
\arguments{
\item{x}{(Optional) Either a URL to the remote MLflow server or the file store,
i.e. the root of the backing file store for experiment and run data. If not
specified, will launch and connect to a local instance listening on a random port.}

\item{activate}{Whether to set the connction as the active connection, defaults to `TRUE`.}

\item{...}{Optional arguments passed to `mlflow_server()`.}
}
\description{
Connect to local or remote MLflow instance.
}
