% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_create_experiment}
\alias{mlflow_create_experiment}
\title{Create Experiment}
\usage{
mlflow_create_experiment(name, artifact_location = NULL,
  activate = TRUE)
}
\arguments{
\item{name}{The name of the experiment to create.}

\item{artifact_location}{Location where all artifacts for this experiment are stored. If
not provided, the remote server will select an appropriate default.}

\item{activate}{Whether to set the created experiment as the active experiment. Defaults to `TRUE`.}
}
\description{
Creates an MLflow experiment.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()

# create local experiment
mlflow_create_experiment("My Experiment")

# create experiment in remote MLflow server
mlflow_set_tracking_uri("http://tracking-server:5000")
mlflow_create_experiment("My Experiment")
}

}
