% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_create_experiment}
\alias{mlflow_create_experiment}
\title{Create Experiment}
\usage{
mlflow_create_experiment(name, artifact_location = NULL, client = NULL)
}
\arguments{
\item{name}{The name of the experiment to create.}

\item{artifact_location}{Location where all artifacts for this experiment are stored. If
not provided, the remote server will select an appropriate default.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Creates an MLflow experiment and returns its id.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
