% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_create_experiment}
\alias{mlflow_create_experiment}
\title{Create Experiment}
\usage{
mlflow_create_experiment(name, artifact_location = NULL, client = NULL)
}
\arguments{
\item{name}{The name of the experiment to create.}

\item{artifact_location}{Location where all artifacts for this experiment are stored. If
not provided, the remote server will select an appropriate default.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Creates an MLflow experiment.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
\seealso{
Other tracking functions: \code{\link{mlflow_delete_experiment}},
  \code{\link{mlflow_delete_run}},
  \code{\link{mlflow_download_artifacts}},
  \code{\link{mlflow_end_run}},
  \code{\link{mlflow_get_experiment}},
  \code{\link{mlflow_get_metric_history}},
  \code{\link{mlflow_get_run}},
  \code{\link{mlflow_list_artifacts}},
  \code{\link{mlflow_list_experiments}},
  \code{\link{mlflow_list_run_infos}},
  \code{\link{mlflow_log_artifact}},
  \code{\link{mlflow_log_batch}},
  \code{\link{mlflow_log_metric}},
  \code{\link{mlflow_log_param}},
  \code{\link{mlflow_rename_experiment}},
  \code{\link{mlflow_restore_experiment}},
  \code{\link{mlflow_restore_run}},
  \code{\link{mlflow_search_runs}},
  \code{\link{mlflow_set_tag}},
  \code{\link{mlflow_start_run}}
}
\concept{tracking functions}
