% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-generics.R
\name{mlflow_create_run}
\alias{mlflow_create_run}
\title{Create Run}
\usage{
mlflow_create_run(experiment_id, user_id = NULL, run_name = NULL,
  source_type = NULL, source_name = NULL, entry_point_name = NULL,
  start_time = NULL, source_version = NULL, tags = NULL,
  client = NULL)
}
\arguments{
\item{experiment_id}{Unique identifier for the associated experiment.}

\item{user_id}{User ID or LDAP for the user executing the run.}

\item{run_name}{Human readable name for run.}

\item{source_type}{Originating source for this run. One of Notebook, Job, Project, Local or Unknown.}

\item{source_name}{String descriptor for source. For example, name or description of the notebook, or job name.}

\item{entry_point_name}{Name of the entry point for the run.}

\item{start_time}{Unix timestamp of when the run started in milliseconds.}

\item{source_version}{Git version of the source code used to create run.}

\item{tags}{Additional metadata for run in key-value pairs.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
reate a new run within an experiment. A run is usually a single execution of a machine learning or data ETL pipeline.
}
\details{
MLflow uses runs to track Param, Metric, and RunTag, associated with a single execution.

When `client` is not specified, these functions attempt to infer the current active client.
}
\seealso{
Other tracking functions: \code{\link{mlflow_client_download_artifacts}},
  \code{\link{mlflow_client_get_experiment_by_name}},
  \code{\link{mlflow_client_log_artifact}},
  \code{\link{mlflow_client_set_terminated}},
  \code{\link{mlflow_create_experiment}},
  \code{\link{mlflow_delete_experiment}},
  \code{\link{mlflow_delete_run}},
  \code{\link{mlflow_get_experiment}},
  \code{\link{mlflow_get_metric_history}},
  \code{\link{mlflow_get_run}},
  \code{\link{mlflow_list_artifacts}},
  \code{\link{mlflow_list_experiments}},
  \code{\link{mlflow_log_batch}},
  \code{\link{mlflow_log_metric}},
  \code{\link{mlflow_log_param}},
  \code{\link{mlflow_restore_experiment}},
  \code{\link{mlflow_restore_run}},
  \code{\link{mlflow_search_runs}},
  \code{\link{mlflow_set_tag}}
}
\concept{tracking functions}
