% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_delete_experiment}
\alias{mlflow_delete_experiment}
\title{Delete Experiment}
\usage{
mlflow_delete_experiment(experiment_id, client = NULL)
}
\arguments{
\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Marks an experiment and associated runs, params, metrics, etc. for deletion. If the
  experiment uses FileStore, artifacts associated with experiment are also deleted.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
