% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_delete_experiment}
\alias{mlflow_delete_experiment}
\title{Delete Experiment}
\usage{
mlflow_delete_experiment(experiment_id, client = NULL)
}
\arguments{
\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Marks an experiment and associated runs, params, metrics, etc. for deletion. If the
  experiment uses FileStore, artifacts associated with experiment are also deleted.
}
