% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{mlflow_delete_run}
\alias{mlflow_delete_run}
\title{Delete a Run}
\usage{
mlflow_delete_run(run_id, client = NULL)
}
\arguments{
\item{run_id}{Run ID.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Deletes the run with the specified ID.
}
