% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow-fluent.R
\name{mlflow_end_run}
\alias{mlflow_end_run}
\title{End a Run}
\usage{
mlflow_end_run(status = "FINISHED")

mlflow_end_run(status = "FINISHED")
}
\arguments{
\item{status}{Updated status of the run. Defaults to `FINISHED`.}

\item{status}{Ending status of the run, defaults to `FINISHED`.}
}
\description{
End an active MLflow run (if there is one).

End the active run.
}
