% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{mlflow_end_run}
\alias{mlflow_end_run}
\title{End Run}
\usage{
mlflow_end_run(status = "FINISHED")
}
\arguments{
\item{status}{Ending status of the run, defaults to `FINISHED`.}
}
\description{
End the active run.
}
