% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{mlflow_end_run}
\alias{mlflow_end_run}
\title{End a Run}
\usage{
mlflow_end_run(status = c("FINISHED", "SCHEDULED", "FAILED", "KILLED"),
  end_time = NULL, run_id = NULL, client = NULL)
}
\arguments{
\item{status}{Updated status of the run. Defaults to `FINISHED`.}

\item{end_time}{Unix timestamp of when the run ended in milliseconds.}

\item{run_id}{Run ID.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Terminates a run. Attempts to end the current active run if `run_id` is not specified.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
