% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_experiment_create}
\alias{mlflow_experiment_create}
\title{Create Experiment}
\usage{
mlflow_experiment_create(mc, name)
}
\arguments{
\item{mc}{The MLflow connection created using \code{mlflow_connect()}.}

\item{name}{The name of the experiment to create.}
}
\description{
Creates an MLflow experiment.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()

mc <- mlflow_connect()
mlflow_experiments(mc)
}

}
