% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_experiments}
\alias{mlflow_experiments}
\title{List Experiments}
\usage{
mlflow_experiments()
}
\description{
Retrieves MLflow experiments as a data frame.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()

# list local experiments
mlflow_experiments(mc)

# list experiments in remote MLflow server
mlflow_tracking_url("http://tracking-server:5000")
mlflow_experiments(mc)
}

}
