% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-generics.R
\name{mlflow_get_experiment}
\alias{mlflow_get_experiment}
\title{Get Experiment}
\usage{
mlflow_get_experiment(experiment_id, client = NULL)
}
\arguments{
\item{experiment_id}{Identifer to get an experiment.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Gets metadata for an experiment and a list of runs for the experiment.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
\seealso{
Other tracking functions: \code{\link{mlflow_client_download_artifacts}},
  \code{\link{mlflow_client_get_experiment_by_name}},
  \code{\link{mlflow_client_list_artifacts}},
  \code{\link{mlflow_client_log_artifact}},
  \code{\link{mlflow_client_log_param}},
  \code{\link{mlflow_client_set_tag}},
  \code{\link{mlflow_client_set_terminated}},
  \code{\link{mlflow_create_experiment}},
  \code{\link{mlflow_create_run}},
  \code{\link{mlflow_delete_experiment}},
  \code{\link{mlflow_delete_run}},
  \code{\link{mlflow_get_run}},
  \code{\link{mlflow_list_experiments}},
  \code{\link{mlflow_log_batch}},
  \code{\link{mlflow_log_metric}},
  \code{\link{mlflow_restore_experiment}},
  \code{\link{mlflow_restore_run}}
}
\concept{tracking functions}
