% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_get_experiment}
\alias{mlflow_get_experiment}
\title{Get Experiment}
\usage{
mlflow_get_experiment(experiment_id = NULL, name = NULL,
  client = NULL)
}
\arguments{
\item{experiment_id}{Identifer to get an experiment.}

\item{name}{The experiment name. Only one of `name` or `experiment_id` should be specified.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Gets metadata for an experiment and a list of runs for the experiment. Attempts to obtain the
active experiment if both `experiment_id` and `name` are unspecified.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
