% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_get_experiment}
\alias{mlflow_get_experiment}
\title{Get Experiment}
\usage{
mlflow_get_experiment(name = NULL, experiment_id = NULL,
  client = NULL)
}
\arguments{
\item{name}{The experiment name, either this or `experiment_id` should be specified.}

\item{experiment_id}{Identifer to get an experiment. Attempts to obtain the active experiment
if not provided.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Gets metadata for an experiment and a list of runs for the experiment.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
