% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_get_experiment}
\alias{mlflow_get_experiment}
\title{Get Experiment}
\usage{
mlflow_get_experiment(experiment_id = NULL, name = NULL, client = NULL)
}
\arguments{
\item{experiment_id}{ID of the experiment.}

\item{name}{The experiment name. Only one of `name` or `experiment_id` should be specified.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Gets metadata for an experiment and a list of runs for the experiment. Attempts to obtain the
active experiment if both `experiment_id` and `name` are unspecified.
}
