% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_get_metric}
\alias{mlflow_get_metric}
\alias{mlflow_get_metric.mlflow_client}
\title{Get Metric}
\usage{
mlflow_get_metric(metric_key, client = NULL, ...)

\method{mlflow_get_metric}{mlflow_client}(metric_key, client = NULL,
  run_id, ...)
}
\arguments{
\item{metric_key}{Name of the metric.}

\item{run_id}{Run ID.}
}
\description{
API to retrieve the logged value for a metric during a run. For a run, if this
  metric is logged more than once, this API will retrieve only the latest value logged.
}
