% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_get_metric}
\alias{mlflow_get_metric}
\title{Get Metric}
\usage{
mlflow_get_metric(metric_key, run_uuid = NULL)
}
\arguments{
\item{metric_key}{Name of the metric.}

\item{run_uuid}{Unique ID for the run for which metric is recorded.}
}
\description{
API to retrieve the logged value for a metric during a run. For a run, if this
  metric is logged more than once, this API will retrieve only the latest value logged.
}
