% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_get_metric_history}
\alias{mlflow_get_metric_history}
\title{Get Metric History}
\usage{
mlflow_get_metric_history(metric_key, client, run_id)
}
\arguments{
\item{metric_key}{Name of the metric.}

\item{client}{An `mlflow_client` object.}

\item{run_id}{Run ID.}
}
\description{
For cases that a metric is logged more than once during a run, this API can be used
  to retrieve all logged values for this metric.
}
