% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_get_metric_history}
\alias{mlflow_get_metric_history}
\title{Get Metric History}
\usage{
mlflow_get_metric_history(metric_key, run_uuid = NULL)
}
\arguments{
\item{metric_key}{Name of the metric.}

\item{run_uuid}{Unique ID for the run for which metric is recorded.}
}
\description{
For cases that a metric is logged more than once during a run, this API can be used
  to retrieve all logged values for this metric.
}
