% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-generics.R
\name{mlflow_get_metric_history}
\alias{mlflow_get_metric_history}
\title{Get Metric History}
\usage{
mlflow_get_metric_history(run_id, metric_key, client = NULL)
}
\arguments{
\item{run_id}{Run ID.}

\item{metric_key}{Name of the metric.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Get a list of all values for the specified metric for a given run.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
\seealso{
Other tracking functions: \code{\link{mlflow_client_download_artifacts}},
  \code{\link{mlflow_client_get_experiment_by_name}},
  \code{\link{mlflow_client_list_artifacts}},
  \code{\link{mlflow_client_log_artifact}},
  \code{\link{mlflow_client_set_terminated}},
  \code{\link{mlflow_create_experiment}},
  \code{\link{mlflow_create_run}},
  \code{\link{mlflow_delete_experiment}},
  \code{\link{mlflow_delete_run}},
  \code{\link{mlflow_get_experiment}},
  \code{\link{mlflow_get_run}},
  \code{\link{mlflow_list_experiments}},
  \code{\link{mlflow_log_batch}},
  \code{\link{mlflow_log_metric}},
  \code{\link{mlflow_log_param}},
  \code{\link{mlflow_restore_experiment}},
  \code{\link{mlflow_restore_run}},
  \code{\link{mlflow_set_tag}}
}
\concept{tracking functions}
