% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{mlflow_get_metric_history}
\alias{mlflow_get_metric_history}
\title{Get Metric History}
\usage{
mlflow_get_metric_history(metric_key, run_id = NULL, client = NULL)
}
\arguments{
\item{metric_key}{Name of the metric.}

\item{run_id}{Run ID.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Get a list of all values for the specified metric for a given run.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
