% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{mlflow_get_metric_history}
\alias{mlflow_get_metric_history}
\title{Get Metric History}
\usage{
mlflow_get_metric_history(metric_key, run_id = NULL, client = NULL)
}
\arguments{
\item{metric_key}{Name of the metric.}

\item{run_id}{Run ID.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Get a list of all values for the specified metric for a given run.
}
