% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{mlflow_get_model_version_by_alias}
\alias{mlflow_get_model_version_by_alias}
\title{Get model version by alias}
\usage{
mlflow_get_model_version_by_alias(name, alias, client = NULL, ...)
}
\arguments{
\item{name}{Name of the registered model.}

\item{alias}{Alias to resolve.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}

\item{...}{Reserved for future options.}
}
\description{
Resolves an alias to a registered model version.
}
\details{
To load the aliased model contents, use
  `mlflow_load_model("models:/<model_name>@<alias>")`.
}
