% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{mlflow_get_param}
\alias{mlflow_get_param}
\alias{mlflow_get_param.mlflow_client}
\title{Get Param}
\usage{
mlflow_get_param(param_name, client = NULL, ...)

\method{mlflow_get_param}{mlflow_client}(param_name, client = NULL,
  run_id, ...)
}
\arguments{
\item{param_name}{Name of the param. This field is required.}
}
\value{
The param value as a named list.
}
\description{
Get a param value.
}
