% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_get_param}
\alias{mlflow_get_param}
\title{Get Param}
\usage{
mlflow_get_param(param_name, run_uuid = NULL)
}
\arguments{
\item{param_name}{Name of the param. This field is required.}

\item{run_uuid}{ID of the run from which to retrieve the param value.}
}
\value{
The param value as a named list.
}
\description{
Get a param value.
}
