% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{mlflow_get_run}
\alias{mlflow_get_run}
\title{Get Run}
\usage{
mlflow_get_run(run_id = NULL, client = NULL)
}
\arguments{
\item{run_id}{Run ID.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Gets metadata, params, tags, and metrics for a run. Returns a single value for each metric
key: the most recently logged metric value at the largest step.
}
