% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{mlflow_install}
\alias{mlflow_install}
\title{Install MLflow}
\usage{
mlflow_install()
}
\description{
Installs MLflow for individual use.
}
\details{
Notice that MLflow requires Python and pip to be installed.
See \url{https://www.python.org/getit/} and \url{https://pip.pypa.io/}.

@export
}
