% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{mlflow_install}
\alias{mlflow_install}
\title{Install MLflow}
\usage{
mlflow_install()
}
\description{
Installs MLflow for individual use.
}
\details{
MLflow requires Python and Conda to be installed.
See \url{https://www.python.org/getit/} and \url{https://docs.conda.io/projects/conda/en/latest/user-guide/install/}.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()
}

}
