% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-rest.R
\name{mlflow_list_experiments}
\alias{mlflow_list_experiments}
\title{List Experiments}
\usage{
mlflow_list_experiments()
}
\description{
Retrieves MLflow experiments as a data frame.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()

# list local experiments
mlflow_list_experiments()

# list experiments in remote MLflow server
mlflow_set_tracking_uri("http://tracking-server:5000")
mlflow_list_experiments()
}

}
