% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_list_experiments}
\alias{mlflow_list_experiments}
\title{List Experiments}
\usage{
mlflow_list_experiments(
  view_type = c("ACTIVE_ONLY", "DELETED_ONLY", "ALL"),
  client = NULL
)
}
\arguments{
\item{view_type}{Qualifier for type of experiments to be returned. Defaults to `ACTIVE_ONLY`.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Gets a list of all experiments.
}
