% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mlflow_load_flavor}
\alias{mlflow_load_flavor}
\title{Load MLflow Model Flavor}
\usage{
mlflow_load_flavor(flavor, model_path)
}
\arguments{
\item{flavor}{An MLflow flavor object.}

\item{model_path}{The path to the MLflow model wrapped in the correct
class.}
}
\description{
Loads an MLflow model using a specific flavor. This method is called internally by
\link[mlflow]{mlflow_load_model}, but is exposed for package authors to extend the supported
MLflow models.
}
