% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-flavor.R
\name{mlflow_load_flavor}
\alias{mlflow_load_flavor}
\title{Load MLflow Model Flavor}
\usage{
mlflow_load_flavor(model_path)
}
\arguments{
\item{model_path}{The path to the MLflow model wrapped in the correct
class.}
}
\description{
Loads an MLflow model flavor, to be used by package authors
to extend the supported MLflow models.
}
