% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mlflow_load_model}
\alias{mlflow_load_model}
\title{Load MLflow Model}
\usage{
mlflow_load_model(model_path, flavor = NULL, run_id = NULL)
}
\arguments{
\item{model_path}{Path to the MLflow model. The path is relative to the run with the given
run-id or local filesystem path without run-id.}

\item{flavor}{Optional flavor specification. Can be used to load a particular flavor in case
there are multiple flavors available.}

\item{run_id}{Optional MLflow run-id. If supplied model will be fetched from MLflow tracking
server.}
}
\description{
Loads an MLflow model. MLflow models can have multiple model flavors. Not all flavors / models
can be loaded in R. This method by default searches for a flavor supported by R/MLflow.
}
