% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_log_artifact}
\alias{mlflow_log_artifact}
\title{Log Artifact}
\usage{
mlflow_log_artifact(path)
}
\arguments{
\item{path}{The file or directory to log as an artifact.}
}
\description{
Logs an specific file or directory as an artifact.
}
