% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{mlflow_log_batch}
\alias{mlflow_log_batch}
\title{Log Batch}
\usage{
mlflow_log_batch(metrics = NULL, params = NULL, tags = NULL,
  timestamps = NULL, run_id = NULL, client = NULL)
}
\arguments{
\item{metrics}{A named list of metrics to log.}

\item{params}{A named list of params to log.}

\item{tags}{A named list of tags to log.}

\item{timestamps}{(Optional) A list of timestamps of the same length as `metrics`.}

\item{run_id}{Run ID.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Log a batch of metrics, params, and/or tags for a run. The server will respond with an error (non-200 status code)
  if any data failed to be persisted. In case of error (due to internal server error or an invalid request), partial
  data may be written.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
\seealso{
Other tracking functions: \code{\link{mlflow_create_experiment}},
  \code{\link{mlflow_delete_experiment}},
  \code{\link{mlflow_delete_run}},
  \code{\link{mlflow_download_artifacts}},
  \code{\link{mlflow_end_run}},
  \code{\link{mlflow_get_experiment}},
  \code{\link{mlflow_get_metric_history}},
  \code{\link{mlflow_get_run}},
  \code{\link{mlflow_list_artifacts}},
  \code{\link{mlflow_list_experiments}},
  \code{\link{mlflow_list_run_infos}},
  \code{\link{mlflow_log_artifact}},
  \code{\link{mlflow_log_metric}},
  \code{\link{mlflow_log_param}},
  \code{\link{mlflow_rename_experiment}},
  \code{\link{mlflow_restore_experiment}},
  \code{\link{mlflow_restore_run}},
  \code{\link{mlflow_search_runs}},
  \code{\link{mlflow_set_tag}},
  \code{\link{mlflow_start_run}}
}
\concept{tracking functions}
