% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{mlflow_log_metric}
\alias{mlflow_log_metric}
\title{Log Metric}
\usage{
mlflow_log_metric(key, value, timestamp = NULL, run_id = NULL,
  client = NULL)
}
\arguments{
\item{key}{Name of the metric.}

\item{value}{Float value for the metric being logged.}

\item{timestamp}{Unix timestamp in milliseconds at the time metric was logged.}

\item{run_id}{Run ID.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Logs a metric for a run. Metrics key-value pair that records a single float measure.
  During a single execution of a run, a particular metric can be logged several times.
  Backend will keep track of historical values along with timestamps.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
