% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mlflow_log_model}
\alias{mlflow_log_model}
\title{Log Model}
\usage{
mlflow_log_model(f, path = "mlflow-model")
}
\arguments{
\item{f}{The serving function that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}
}
\description{
Logs a model in the given run. Similar to `mlflow_save_model()`
but stores model only as an artifact within the active run.
}
