% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{mlflow_log_model}
\alias{mlflow_log_model}
\title{Log Model}
\usage{
mlflow_log_model(fn, artifact_path, run_uuid = NULL)
}
\arguments{
\item{fn}{The serving function that will perform a prediction.}

\item{artifact_path}{Destination path where this MLflow compatible model
will be saved.}

\item{run_uuid}{The run associated with the model to be logged.}
}
\description{
Logs a model in the given run. Similar to `mlflow_save_model()`
but stores model as an artifact within the active run.
}
