% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mlflow_log_model}
\alias{mlflow_log_model}
\title{Log Model}
\usage{
mlflow_log_model(fn, artifact_path)
}
\arguments{
\item{fn}{The serving function that will perform a prediction.}

\item{artifact_path}{Destination path where this MLflow compatible model
will be saved.}
}
\description{
Logs a model for this run. Similar to `mlflow_save_model()`
but stores model as an artifact within the active run.
}
