% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R, R/tracking-fluent.R
\name{mlflow_log_param}
\alias{mlflow_log_param}
\alias{mlflow_log_param.mlflow_client}
\alias{mlflow_log_param.NULL}
\title{Log Parameter}
\usage{
mlflow_log_param(key, value, client = NULL, ...)

\method{mlflow_log_param}{mlflow_client}(key, value, client = NULL,
  run_id, ...)

\method{mlflow_log_param}{NULL}(key, value, client = NULL, ...)
}
\arguments{
\item{key}{Name of the parameter.}

\item{value}{String value of the parameter.}

\item{run_id}{Run ID.}
}
\description{
API to log a parameter used for this run. Examples are params and hyperparams
  used for ML training, or constant dates and values used in an ETL pipeline.
  A params is a STRING key-value pair. For a run, a single parameter is allowed
  to be logged only once.
}
