% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-fluent.R
\name{mlflow_log_param}
\alias{mlflow_log_param}
\title{Log Parameter}
\usage{
mlflow_log_param(key, value)
}
\arguments{
\item{key}{Name of the parameter.}

\item{value}{String value of the parameter.}
}
\description{
API to log a parameter used for this run. Examples are params and hyperparams
  used for ML training, or constant dates and values used in an ETL pipeline.
  A params is a STRING key-value pair. For a run, a single parameter is allowed
  to be logged only once.
}
\details{
The fluent API family of functions operate with an implied MLflow client
  determined by the service set by `mlflow_set_tracking_uri()`. For operations
  involving a run it adopts the current active run, or, if one does not exist,
  starts one through the implied service.
}
