% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/param.R
\name{mlflow_param}
\alias{mlflow_param}
\title{Read Command Line Parameter}
\usage{
mlflow_param(name, default = NULL, description = NULL)
}
\arguments{
\item{name}{The name for this parameter.}

\item{default}{The default value for this parameter.}

\item{description}{Optional description for this parameter.}
}
\description{
Reads a command line parameter.
}
