% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mlflow_predict}
\alias{mlflow_predict}
\title{Predict using MLflow Model}
\usage{
mlflow_predict(model_dir, data, output_file = NULL, restore = FALSE)
}
\arguments{
\item{model_dir}{The path to the MLflow model, as a string.}

\item{data}{Data frame, 'JSON' or 'CSV' file to be used for prediction.}

\item{output_file}{'JSON' or 'CSV' file where the prediction will be written to.}

\item{restore}{Should \code{mlflow_restore()} be called before serving?}
}
\description{
Predict using a MLflow Model from a 'JSON' file.
}
\examples{
\dontrun{
library(mlflow)

# save simple model which roundtrips data as prediction
mlflow_save_model(function(df) df, "mlflow_roundtrip")

# save data as json
jsonlite::write_json(iris, "iris.json")

# serve an existing model over a web interface
mlflow_predict("mlflow_roundtrip", "iris.json")
}

}
