% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mlflow_predict_model}
\alias{mlflow_predict_model}
\title{Generate Prediction with MLflow Model}
\usage{
mlflow_predict_model(model, data)
}
\arguments{
\item{model}{MLflow model.}

\item{data}{Dataframe to be scored.}
}
\description{
Generates a prediction with an MLflow model.
}
