% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_rename_experiment}
\alias{mlflow_rename_experiment}
\title{Rename Experiment}
\usage{
mlflow_rename_experiment(new_name, experiment_id = NULL, client = NULL)
}
\arguments{
\item{new_name}{The experiment’s name will be changed to this. The new name must be unique.}

\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Renames an experiment.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
