% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_rename_experiment}
\alias{mlflow_rename_experiment}
\title{Rename Experiment}
\usage{
mlflow_rename_experiment(new_name, experiment_id = NULL, client = NULL)
}
\arguments{
\item{new_name}{The experiment's name will be changed to this. The new name must be unique.}

\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Renames an experiment.
}
