% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-generics.R
\name{mlflow_restore_experiment}
\alias{mlflow_restore_experiment}
\title{Restore Experiment}
\usage{
mlflow_restore_experiment(experiment_id, client = NULL)
}
\arguments{
\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Restores an experiment marked for deletion. This also restores associated metadata,
  runs, metrics, and params. If experiment uses FileStore, underlying artifacts
  associated with experiment are also restored.
}
\details{
Throws `RESOURCE_DOES_NOT_EXIST` if the experiment was never created or was permanently deleted.

When `client` is not specified, these functions attempt to infer the current active client.
}
\seealso{
Other tracking functions: \code{\link{mlflow_client_download_artifacts}},
  \code{\link{mlflow_client_get_experiment_by_name}},
  \code{\link{mlflow_client_list_artifacts}},
  \code{\link{mlflow_client_log_artifact}},
  \code{\link{mlflow_client_log_param}},
  \code{\link{mlflow_client_set_terminated}},
  \code{\link{mlflow_create_experiment}},
  \code{\link{mlflow_create_run}},
  \code{\link{mlflow_delete_experiment}},
  \code{\link{mlflow_delete_run}},
  \code{\link{mlflow_get_experiment}},
  \code{\link{mlflow_get_run}},
  \code{\link{mlflow_list_experiments}},
  \code{\link{mlflow_log_batch}},
  \code{\link{mlflow_log_metric}},
  \code{\link{mlflow_restore_run}},
  \code{\link{mlflow_set_tag}}
}
\concept{tracking functions}
