% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/run.R
\name{mlflow_run}
\alias{mlflow_run}
\title{Run in MLflow}
\usage{
mlflow_run(x)
}
\arguments{
\item{x}{A file, expression or function to run.}
}
\description{
Runs the given file, expression or function within
the context of an MLflow run.
}
