% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{mlflow_run}
\alias{mlflow_run}
\title{Run in MLflow}
\usage{
mlflow_run(uri, entry_point = NULL, param_list = NULL,
  experiment_id = NULL)
}
\arguments{
\item{uri}{A directory or an R script.}

\item{entry_point}{Entry point within project, defaults to `main` if not specified.}

\item{param_list}{A list of parameters.}

\item{experiment_id}{ID of the experiment under which to launch the run.}
}
\description{
Runs the given file, expression or function within
the context of an MLflow run.
}
