% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-run.R
\name{mlflow_run}
\alias{mlflow_run}
\title{Run an MLflow Project}
\usage{
mlflow_run(entry_point = NULL, uri = ".", version = NULL,
  param_list = NULL, experiment_id = NULL, experiment_name = NULL,
  backend = NULL, backend_config = NULL, no_conda = FALSE,
  storage_dir = NULL)
}
\arguments{
\item{entry_point}{Entry point within project, defaults to `main` if not specified.}

\item{uri}{A directory containing modeling scripts, defaults to the current directory.}

\item{version}{Version of the project to run, as a Git commit reference for Git projects.}

\item{param_list}{A list of parameters.}

\item{experiment_id}{ID of the experiment under which to launch the run.}

\item{experiment_name}{Name of the experiment under which to launch the run.}

\item{backend}{Execution backend to use for run.}

\item{backend_config}{Path to JSON file which will be passed to the backend. For the Databricks backend,
it should describe the cluster to use when launching a run on Databricks.}

\item{no_conda}{If specified, assume that MLflow is running within a Conda environment with the necessary
dependencies for the current project instead of attempting to create a new Conda environment. Only
valid if running locally.}

\item{storage_dir}{Valid only when `backend` is local. MLflow downloads artifacts from distributed URIs passed to
parameters of type `path` to subdirectories of `storage_dir`.}
}
\value{
The run associated with this run.
}
\description{
Wrapper for `mlflow run`.
}
