% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-flavor.R
\name{mlflow_save_flavor}
\alias{mlflow_save_flavor}
\title{Save MLflow Model Flavor}
\usage{
mlflow_save_flavor(x, path = "model", r_dependencies = NULL,
  conda_env = NULL)
}
\arguments{
\item{x}{The serving function or model that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}

\item{r_dependencies}{Optional vector of paths to dependency files
to include in the model, as in \code{r-dependencies.txt}
or \code{conda.yaml}.}

\item{conda_env}{Path to Conda dependencies file.}
}
\value{
This function must return a list of flavors that conform to
  the MLmodel specification.
}
\description{
Saves model in MLflow flavor, to be used by package authors
to extend the supported MLflow models.
}
