% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-flavor.R
\name{mlflow_save_flavor}
\alias{mlflow_save_flavor}
\title{Save MLflow Model Flavor}
\usage{
mlflow_save_flavor(model, path = "model", ...)
}
\arguments{
\item{model}{The model that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}

\item{...}{Optional additional arguments passed to `mlflow_save_flavor()` - for example,
`conda_env = /path/to/conda.yaml` may be passed to specify a conda dependencies file
for flavors (e.g. keras) that support conda environments.}
}
\value{
This function must return a list of flavors that conform to
  the MLmodel specification.
}
\description{
Saves model in MLflow flavor, to be used by package authors
to extend the supported MLflow models.
}
