% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mlflow_save_model}
\alias{mlflow_save_model}
\title{Save Model for MLflow}
\usage{
mlflow_save_model(model, path = "model", ...)
}
\arguments{
\item{model}{The model that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}

\item{...}{Optional additional arguments passed to `mlflow_save_flavor()` - for example,
`conda_env = /path/to/conda.yaml` may be passed to specify a conda dependencies file
for flavors (e.g. keras) that support conda environments.}
}
\description{
Saves model in MLflow format that can later be used
for prediction and serving.
}
