% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-crate.R, R/model-keras.R, R/model.R
\name{mlflow_save_model.crate}
\alias{mlflow_save_model.crate}
\alias{mlflow_save_model.keras.engine.training.Model}
\alias{mlflow_save_model}
\title{Save Model for MLflow}
\usage{
\method{mlflow_save_model}{crate}(model, path, ...)

\method{mlflow_save_model}{keras.engine.training.Model}(model, path, conda_env = NULL, ...)

mlflow_save_model(model, path, ...)
}
\arguments{
\item{model}{The model that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}

\item{...}{Optional additional arguments.}

\item{conda_env}{Path to Conda dependencies file.}
}
\description{
Saves model in MLflow format that can later be used for prediction and serving. This method is
generic to allow package authors to save custom model types.
}
