% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mlflow_save_model}
\alias{mlflow_save_model}
\title{Save Model for MLflow}
\usage{
mlflow_save_model(x, path = "model", r_dependencies = NULL,
  conda_env = NULL)
}
\arguments{
\item{x}{The serving function or model that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}

\item{r_dependencies}{Optional vector of paths to dependency files
to include in the model, as in \code{r-dependencies.txt}
or \code{conda.yaml}.}

\item{conda_env}{Path to Conda dependencies file.}
}
\description{
Saves model in MLflow format that can later be used
for prediction and serving.
}
