% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_set_experiment}
\alias{mlflow_set_experiment}
\title{Set Experiment}
\usage{
mlflow_set_experiment(experiment_name = NULL, experiment_id = NULL,
  artifact_location = NULL)
}
\arguments{
\item{experiment_name}{Name of experiment to be activated.}

\item{experiment_id}{ID of experiment to be activated.}

\item{artifact_location}{Location where all artifacts for this experiment are stored. If
not provided, the remote server will select an appropriate default.}
}
\description{
Sets an experiment as the active experiment. If the experiment does not exist,
  creates an experiment with provided name.
}
\details{
The fluent API family of functions operate with an implied MLflow client
  determined by the service set by `mlflow_set_tracking_uri()`. For operations
  involving a run it adopts the current active run, or, if one does not exist,
  starts one through the implied service.
}
