% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_set_experiment}
\alias{mlflow_set_experiment}
\title{Set Experiment}
\usage{
mlflow_set_experiment(
  experiment_name = NULL,
  experiment_id = NULL,
  artifact_location = NULL
)
}
\arguments{
\item{experiment_name}{Name of experiment to be activated.}

\item{experiment_id}{ID of experiment to be activated.}

\item{artifact_location}{Location where all artifacts for this experiment are stored. If
not provided, the remote server will select an appropriate default.}
}
\description{
Sets an experiment as the active experiment. Either the name or ID of the experiment can be provided.
  If the a name is provided but the experiment does not exist, this function creates an experiment
  with provided name. Returns the ID of the active experiment.
}
