% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{mlflow_set_experiment_tag}
\alias{mlflow_set_experiment_tag}
\title{Set Experiment Tag}
\usage{
mlflow_set_experiment_tag(key, value, experiment_id = NULL, client = NULL)
}
\arguments{
\item{key}{Name of the tag. All storage backends are guaranteed to support
key values up to 250 bytes in size. This field is required.}

\item{value}{String value of the tag being logged. All storage backends are
guaranteed to support key values up to 5000 bytes in size. This field is required.}

\item{experiment_id}{ID of the experiment.}

\item{client}{(Optional) An MLflow client object returned from \link[mlflow]{mlflow_client}.
If specified, MLflow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
MLflow will use the tracking server associated with the current tracking URI.}
}
\description{
Sets a tag on an experiment with the specified ID. Tags are experiment metadata that can be updated.
}
