% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{mlflow_set_tag}
\alias{mlflow_set_tag}
\title{Set Tag}
\usage{
mlflow_set_tag(key, value, run_id = NULL, client = NULL)
}
\arguments{
\item{key}{Name of the tag. Maximum size is 255 bytes. This field is required.}

\item{value}{String value of the tag being logged. Maximum size is 500 bytes. This field is required.}

\item{run_id}{Run ID.}

\item{client}{(Optional) An `mlflow_client` object.}
}
\description{
Sets a tag on a run. Tags are run metadata that can be updated during a run and
 after a run completes.
}
\details{
When `client` is not specified, these functions attempt to infer the current active client.
}
\seealso{
Other tracking functions: \code{\link{mlflow_create_experiment}},
  \code{\link{mlflow_delete_experiment}},
  \code{\link{mlflow_delete_run}},
  \code{\link{mlflow_download_artifacts}},
  \code{\link{mlflow_end_run}},
  \code{\link{mlflow_get_experiment}},
  \code{\link{mlflow_get_metric_history}},
  \code{\link{mlflow_get_run}},
  \code{\link{mlflow_list_artifacts}},
  \code{\link{mlflow_list_experiments}},
  \code{\link{mlflow_list_run_infos}},
  \code{\link{mlflow_log_artifact}},
  \code{\link{mlflow_log_batch}},
  \code{\link{mlflow_log_metric}},
  \code{\link{mlflow_log_param}},
  \code{\link{mlflow_rename_experiment}},
  \code{\link{mlflow_restore_experiment}},
  \code{\link{mlflow_restore_run}},
  \code{\link{mlflow_search_runs}},
  \code{\link{mlflow_start_run}}
}
\concept{tracking functions}
