% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-source.R
\name{mlflow_source}
\alias{mlflow_source}
\title{Source a Script with MLflow Params}
\usage{
mlflow_source(uri)
}
\arguments{
\item{uri}{Path to an R script, can be a quoted or unquoted string.}
}
\description{
This function should not be used interactively. It is designed to be called via `Rscript` from
  the terminal or through the MLflow CLI.
}
\keyword{internal}
