% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{mlflow_ui}
\alias{mlflow_ui}
\title{MLflow User Interface}
\usage{
mlflow_ui(x, ...)
}
\arguments{
\item{x}{If specified, can be either an `mlflow_connection` object or a string
specifying the file store, i.e. the root of the backing file store for experiment
and run data.}

\item{...}{Optional arguments passed to `mlflow_server()` when `x` is a path to a file store.}
}
\description{
Launches MLflow user interface.
}
\examples{
\dontrun{
library(mlflow)
mlflow_install()

# launch mlflow ui locally
mlflow_ui()

# launch mlflow ui for existing mlflow server
mlflow_set_tracking_uri("http://tracking-server:5000")
mlflow_ui()
}

}
