% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%||\%}
\alias{\%>\%}
\alias{\%<-\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{purrr}{\code{\link[purrr:pipe]{\%>\%}}}

  \item{rlang}{\code{\link[rlang:op-null-default]{\%||\%}}}

  \item{zeallot}{\code{\link[zeallot:operator]{\%<-\%}}}
}}

